+++ CHANGES.md
@@ -11,6 +11,11 @@
 
 ## develop
 
+- [CHANGE] `Sora.Config` のフィールド名および `Sora.VideoCodec`, `Sora.AudioCodec` の名前を変更しました。本バージョンを利用する際には以下のように名前を変更してください。
+  - `VideoCodec` → `VideoCodecType`
+  - `VideoBitrate` → `VideoBitRate`
+  - `AudioCodec` → `AudioCodecType`
+  - `AudioBitrate` → `AudioBitRate`
 - [ADD] プッシュ通知に対応
     - @melpon